tests unitarios

+++ src/objeto.rs
@@ -1,9 +1,9 @@
 use crate::direccion::Direccion;
 use std::collections::HashSet;
 
-#[derive(Clone)]
+#[derive(PartialEq, Debug, Clone)]
 pub enum Objeto {
-    Enemigo(i32, HashSet<(usize, usize)>),
+    Enemigo(i32, HashSet<(i32, i32)>),
     Bomba(bool, i32),
     Roca,
     Pared,
